requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
openai>=1.0.0
python-dotenv>=1.0.0

# PyTorch installation:
# For GPU (CUDA 12.1): pip install -r requirements-cuda.txt
# For CPU-only: pip install torch torchvision torchaudio
